Compile & warning fixes
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22346 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/software/clip_znear.h
@@ -29,7 +29,7 @@ class ClipMeatZNear
 {
   int width2;
   int height2;
-  int aspect;
+  float aspect;
   const csVector3* persp;
 
   CS_FORCEINLINE void Project (csVector3& v, const float com_iz)
